refactor: separate functionality
